reorganize the dnn/layers
