+++ brainpy/dnn/layers/__init__.py
@@ -1,3 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .layers import *
+from .activation import *
+from .conv import *
+from .dropout import *
+from .linear import *
